Updated removeEvent to include user id

+++ Back/removeEvent.php
@@ -4,14 +4,17 @@
 
 	if (isset($_POST['event_id'])) {
 
-	$sql = "DELETE FROM EVENTS WHERE event_id = " . $_POST["event_id"] . ";";
+	$sql = "DELETE FROM EVENTS WHERE event_id = " . $_POST["event_id"] . " AND user_id = " . $_POST["user_id"] . ";";
 
 	$response = array(
 			"sql" => $sql
 		);
 
-
-		if (mysqli_query($link, $sql) && mysql_affected_rows() > 0) {
+		//The first if statement on line 14 can be split into two more ifs.
+		//if -> query
+		//    |-> if affected > 0 => event deleted
+		//	  |-> if affected < 0 => event doesnt exist anymore
+		if (mysqli_query($link, $sql) ) {
 			// echo "Event updated successfully";
 			$response = array( "status" => 200,
 				"msg" => "Event Deleted!",
